requirements.txt
google_api_python_client==2.147.0
numpy==1.23.5
pygame==2.6.0
python-dotenv==1.0.1
python_vlc==3.0.20123
yt_dlp==2024.10.22
